better directory structyre
