+++ apps/options-board/components/Board/Board.js
@@ -1,56 +1,21 @@
-export const SVG_NS_URI = 'http://www.w3.org/2000/svg'
+import { getMousePosition } from 'board/utils/svg'
+import createPostIt from './createPostIt'
+import { getPostItColor } from './colors'
 
 const TEMPLATE = `
     <svg viewBox="0 0 1000 1000" version="1.1" xmlns="http://www.w3.org/2000/svg">
     </svg>
 `
 
+const DRAGGABLE_CLASS = 'draggable'
+
 const DEFAULT_POST_IT_ATTRS = {
-  class: 'draggable',
+  class: DRAGGABLE_CLASS,
   width: 127,
   height: 76
 }
 
-const createPostIt = ({ x, y, color, index, attrs = {} }) => {
-  console.log(x, y)
-  const group = document.createElementNS(SVG_NS_URI, 'g')
-
-  const postIt = document.createElementNS(SVG_NS_URI, 'rect')
-  postIt.setAttribute('x', x)
-  postIt.setAttribute('y', y)
-  postIt.setAttribute('fill', color)
-  postIt.setAttribute('data-index', index)
-
-  const currentAttrs = {
-    ...DEFAULT_POST_IT_ATTRS,
-    ...attrs
-  }
-
-  Object.keys(currentAttrs).forEach(attributeName => {
-    postIt.setAttribute(attributeName, currentAttrs[attributeName])
-  })
-
-  group.appendChild(postIt)
-
-  return group
-}
-
-const COLORS = [
-  '#E4EF49',
-  '#19A1DF',
-  '#FD43A7',
-  '#FBA435',
-  '#81D0CF',
-  '#FFD333'
-]
-
-const getMousePosition = (svg, event) => {
-  var CTM = svg.getScreenCTM()
-  return {
-    x: Math.floor((event.clientX - CTM.e) / CTM.a),
-    y: Math.floor((event.clientY - CTM.f) / CTM.d)
-  }
-}
+const isDraggable = element => element.classList.contains(DRAGGABLE_CLASS)
 
 class Board extends HTMLElement {
   constructor () {
@@ -81,18 +46,21 @@ class Board extends HTMLElement {
   }
 
   startDrag (event) {
-    if (event.target.classList.contains('draggable')) {
-      this.currentDraggable = event.target
+    const { target } = event
+    if (isDraggable(target)) {
       const offset = getMousePosition(this.svg, event)
-      offset.x -= parseFloat(this.currentDraggable.getAttributeNS(null, 'x'))
-      offset.y -= parseFloat(this.currentDraggable.getAttributeNS(null, 'y'))
+
+      offset.x -= parseFloat(target.getAttributeNS(null, 'x'))
+      offset.y -= parseFloat(target.getAttributeNS(null, 'y'))
+
       this.currentOffset = offset
+      this.currentDraggable = event.target
     }
   }
 
   onDrag (event) {
+    event.preventDefault()
     if (this.currentDraggable) {
-      event.preventDefault()
       const { x, y } = getMousePosition(this.svg, event)
       this.currentDraggable.setAttributeNS(null, 'x', x - this.currentOffset.x)
       this.currentDraggable.setAttributeNS(null, 'y', y - this.currentOffset.y)
@@ -101,7 +69,6 @@ class Board extends HTMLElement {
 
   endDrag (event) {
     if (this.currentDraggable) {
-      console.log(this.data)
       const newData = [...this.data]
       const currentIndex = parseInt(
         this.currentDraggable.getAttribute('data-index')
@@ -116,8 +83,6 @@ class Board extends HTMLElement {
       }
 
       this.data = newData
-
-      console.log(this.data)
     }
 
     this.currentDraggable = undefined
@@ -139,7 +104,8 @@ class Board extends HTMLElement {
         createPostIt({
           ...element,
           index,
-          color: COLORS[index % COLORS.length]
+          color: getPostItColor(index),
+          attrs: DEFAULT_POST_IT_ATTRS
         })
       )
       .forEach(postIt => {
